sql/row: move the fkInsertHelper to its own file
Release note: None
